CellBender Feature Differences

Get quick values for raw counts, CellBender counts, count differences, and percent count differences per feature.

CellBender_Feature_Diff(seurat_object, raw_assay, cell_bender_assay)

Arguments

seurat_object

Seurat object name.

raw_assay

Name of the assay containing the raw count data.

cell_bender_assay

Name of the assay containing the CellBender count data.

Value

A data.frame containing summed raw counts, CellBender counts, count difference, and percent difference in counts.

Examples

if (FALSE) {
cb_stats <- CellBender_Feature_Diff(seurat_object - obj, raw_assay = "RAW",
cell_bender_assay = "RNA")
}