Clustered DotPlot

Clustered DotPlots using ComplexHeatmap

Clustered_DotPlot(
  seurat_object,
  features,
  colors_use_exp = viridis_plasma_dark_high,
  exp_color_min = -2,
  exp_color_middle = NULL,
  exp_color_max = 2,
  print_exp_quantiles = FALSE,
  colors_use_idents = NULL,
  x_lab_rotate = TRUE,
  k = 1,
  row_km_repeats = 1000,
  column_km_repeats = 1000,
  row_label_size = 8,
  raster = FALSE,
  plot_km_elbow = TRUE,
  elbow_kmax = NULL,
  assay = NULL,
  group.by = NULL,
  idents = NULL,
  show_parent_dend_line = TRUE,
  ggplot_default_colors = FALSE,
  color_seed = 123,
  seed = 123
)

Arguments

seurat_object

Seurat object name.

features

Features to plot.

colors_use_exp

Color palette to use for plotting expression scale. Default is viridis::plasma(n = 20, direction = -1).

exp_color_min

Minimum scaled average expression threshold (everything smaller will be set to this). Default is -2.

exp_color_middle

What scaled expression value to use for the middle of the provided colors_use_exp. By default will be set to value in middle of exp_color_min and exp_color_max.

exp_color_max

Minimum scaled average expression threshold (everything smaller will be set to this). Default is 2.

print_exp_quantiles

Whether to print the quantiles of expression data in addition to plots. Default is FALSE. NOTE: These values will be altered by choices of exp_color_min and exp_color_min if there are values below or above those cutoffs, respectively.

colors_use_idents

specify color palette to used for identity labels. By default if number of levels plotted is less than or equal to 36 it will use "polychrome" and if greater than 36 will use "varibow" with shuffle = TRUE both from DiscretePalette_scCustomize.

x_lab_rotate

How to rotate column labels. By default set to TRUE which rotates labels 45 degrees. If set FALSE rotation is set to 0 degrees. Users can also supply custom angle for text rotation.

k

Value to use for k-means clustering on rows. Sets (km) parameter in ComplexHeatmap::Heatmap(). From ComplexHeatmap::Heatmap(): Apply k-means clustering on rows. If the value is larger than 1, the heatmap will be split by rows according to the k-means clustering. For each row slice, hierarchical clustering is still applied with parameters above.

row_km_repeats

Number of k-means runs to get a consensus k-means clustering. Note if row_km_repeats is set to more than one, the final number of groups might be smaller than row_km, but this might mean the original row_km is not a good choice. Default is 1000.

column_km_repeats

Number of k-means runs to get a consensus k-means clustering. Similar as row_km_repeats. Default is 100.

row_label_size

Size of the feature labels. Provided to row_names_gp in Heatmap call.

raster

Logical, whether to render in raster format (faster plotting, smaller files). Default is FALSE.

plot_km_elbow

Logical, whether or not to return the Sum Squared Error Elbow Plot for k-means clustering. Estimating elbow of this plot is one way to determine "optimal" value for k. Based on: https://stackoverflow.com/a/15376462/15568251.

elbow_kmax

The maximum value of k to use for plot_km_elbow. Suggest setting larger value so the true shape of plot can be observed. Value must be 1 less than number of features provided. If NULL parameter will be set dependent on length of feature list up to elbow_kmax = 20.

assay

Name of assay to use, defaults to the active assay.

group.by

Group (color) cells in different ways (for example, orig.ident).

idents

Which classes to include in the plot (default is all).

show_parent_dend_line

Logical, Sets parameter of same name in ComplexHeatmap::Heatmap(). From ComplexHeatmap::Heatmap(): When heatmap is split, whether to add a dashed line to mark parent dendrogram and children dendrograms. Default is TRUE.

ggplot_default_colors

logical. If colors_use = NULL, Whether or not to return plot using default ggplot2 "hue" palette instead of default "polychrome" or "varibow" palettes.

color_seed

random seed for the "varibow" palette shuffle if colors_use = NULL and number of groups plotted is greater than 36. Default = 123.

seed

Sets seed for reproducible plotting (ComplexHeatmap plot).

Value

A ComplexHeatmap or if plot_km_elbow = TRUE a list containing ggplot2 object and ComplexHeatmap.

References

https://divingintogeneticsandgenomics.rbind.io/post/clustered-dotplot-for-single-cell-rnaseq/

See also

Author

Ming Tang (Original Code), Sam Marsh (Wrap single function, added/modified functionality)

Examples

if (FALSE) {
Clustered_DotPlot(seurat_object = object, features = gene_list)
}