Create Seurat Object with Cell Bender and Raw data

Enables easy creation of Seurat object which contains both cell bender data and raw count data as separate assays within the object.

Create_CellBender_Merged_Seurat(
  raw_cell_bender_matrix = NULL,
  raw_counts_matrix = NULL,
  raw_assay_name = "RAW",
  min_cells = 5,
  min_features = 200,
  ...
)

Arguments

raw_cell_bender_matrix
matrix file containing the cell bender correct counts.

raw_counts_matrix
matrix file contain the uncorrected Cell Ranger (or other) counts.

raw_assay_name
a key value to use specifying the name of assay. Default is "RAW".

min_cells
value to supply to min.cells parameter of CreateSeuratObject. Default is 5.

min_features
value to supply to min.features parameter of CreateSeuratObject. Default is 200.

...
Extra parameters passed to CreateSeuratObject.

Value

A Seurat Object contain both the Cell Bender corrected counts ("RNA" assay) and uncorrected counts ("RAW" assay; or other name specified to raw_assay_name).

Examples

if (FALSE) {
seurat_obj <- Create_CellBender_Merged_Seurat(raw_cell_bender_matrix = cb_matrix, raw_counts_matrix = cr_matrix)
}