Customized DotPlot

Code for creating customized DotPlot

Usage

DotPlot_scCustom(
  seurat_object,
  features,
  colors_use = viridis_plasma_dark_high,
  remove_axis_titles = TRUE,
  x_lab_rotate = FALSE,
  y_lab_rotate = FALSE,
  facet_label_rotate = FALSE,
  flip_axes = FALSE,
  ...
)

Arguments

seurat_object

Seurat object name.

features

Features to plot.

colors_use

specify color palette to used. Default is viridis_plasma_dark_high.

remove_axis_titles

logical. Whether to remove the x and y axis titles. Default = TRUE.

x_lab_rotate

Rotate x-axis labels 45 degrees (Default is FALSE).

y_lab_rotate

Rotate x-axis labels 45 degrees (Default is FALSE).

facet_label_rotate

Rotate facet labels on grouped DotPlots by 45 degrees (Default is FALSE).

flip_axes

whether or not to flip and X and Y axes (Default is FALSE).

...

Extra parameters passed to DotPlot.

Value

A ggplot object

Examples

if (FALSE) {
DotPlot_scCustom(seurat_object = object, features = gene_list)
}