Extract sample level meta.data

Returns a by identity meta.data data.frame with one row per sample. Useful for downstream quick view of sample breakdown, meta data table creation, and/or use in pseudobulk analysis

Extract_Sample_Meta(
  object,
  sample_col = "orig.ident",
  sample_name = deprecated(),
  variables_include = NULL,
  variables_exclude = NULL,
  include_all = FALSE
)

Arguments

object

Seurat or LIGER object

sample_col

meta.data column to use as sample. Output data.frame will contain one row per level or unique value in this variable.

sample_name

. See sample_name.

variables_include

@meta.data columns to keep in final data.frame. All other columns will be discarded. Default is NULL.

variables_exclude

columns to discard in final data.frame. Many cell level columns are irrelevant at the sample level (e.g., nFeature_RNA, percent_mito).

Default parameter value is NULL but internally will set to discard nFeature_ASSAY(s), nCount_ASSAY(s), percent_mito, percent_ribo, percent_mito_ribo, and log10GenesPerUMI.

If sample level median values are desired for these type of variables the output of this function can be joined with output of Median_Stats.

Set parameter to include_all = TRUE to prevent any columns from being excluded.

include_all

logical, whether or not to include all object meta data columns in output data.frame. Default is FALSE.

Value

Returns a data.frame with one row per sample_name.

Examples

library(Seurat)
pbmc_small[["batch"]] <- sample(c("batch1", "batch2"), size = ncol(pbmc_small), replace = TRUE)

sample_meta <- Extract_Sample_Meta(object = pbmc_small, sample_name = "orig.ident")
#> Warning: The `sample_name` argument of `Extract_Sample_Meta()` is deprecated as of
#> scCustomize 3.2.1.
#> ℹ The {.code sample_name} parameter is soft-deprecated.  Please update code to
#>   use `sample_col` instead.

# Only return specific columns from meta data (orig.ident and batch)
sample_meta2 <- Extract_Sample_Meta(object = pbmc_small, sample_name = "orig.ident",
variables_include = "batch")

# Return all columns from meta data
sample_meta3 <- Extract_Sample_Meta(object = pbmc_small, sample_name = "orig.ident",
include_all = TRUE)