Factor Correlation Plot

Plot positive correlations between gene loadings across W factor matrix in LIGER or feature loadings in reduction slot of Seurat object.

Factor_Cor_Plot(
  object,
  reduction = NULL,
  colors_use = NULL,
  label = FALSE,
  label_threshold = 0.5,
  label_size = 5,
  plot_title = NULL,
  plot_type = "full",
  positive_only = FALSE,
  x_lab_rotate = TRUE,
  cluster = TRUE,
  cluster_rect = FALSE,
  cluster_rect_num = NULL,
  cluster_rect_col = NULL
)

Arguments

object

LIGER or Seurat object.

reduction

Seurat ONLY; name of dimensionality reduction containing NMF loadings.

colors_use

Color palette to use for correlation values. Default is RColorBrewer::RdBu if positive_only = FALSE. If positive_only = TRUE the default is viridis. Users can also supply vector of 3 colors (low, mid, high).

label

logical, whether to add correlation values to plot result.

label_threshold

threshold for adding correlation values if label = TRUE. Default is 0.5.

label_size

size of correlation labels

plot_title

Plot title.

plot_type

Controls plotting full matrix, or just the upper or lower triangles. Accepted values are: "full" (default), "upper", or "lower".

positive_only

logical, whether to limit the plotted values to only positive correlations (negative values set to 0); default is FALSE.

x_lab_rotate

logical, whether to rotate the axes labels on the x-axis. Default is TRUE.

cluster

logical, whether to cluster the plot using hclust (default TRUE). If FALSE factors are listed in numerical order.

cluster_rect

logical, whether to add rectangles around the clustered areas on plot, default is FALSE. Uses cutree to create groups.

cluster_rect_num

number of rectangles to add to the plot, default NULL. Value is provided to k in cutree.

cluster_rect_col

color to use for rectangles, default MULL (will set color automatically).

Value

A ggplot object

Examples

if (FALSE) {
Factor_Cor_Plot(object = obj)
}