Customize FeaturePlot

Create Custom FeaturePlots and preserve scale (no binning)

Usage

FeaturePlot_scCustom(
  seurat_object,
  features,
  colors_use = viridis_plasma_dark_high,
  na_color = "lightgray",
  order = TRUE,
  pt.size = NULL,
  reduction = NULL,
  na_cutoff = 1e-09,
  raster = NULL,
  split.by = NULL,
  num_columns = NULL,
  slot = "data",
  ...
)

Arguments

seurat_object

Seurat object name.

features

Feature(s) to plot.

colors_use

list of colors or color palette to use.

order

whether to move positive cells to the top (default = TRUE).

pt.size

Adjust point size for plotting.

reduction

Dimensionality Reduction to use (if NULL then defaults to Object default).

na_cutoff

Value to use as minimum expression cutoff. This will be lowest value plotted use palette provided to colors_use. Leave as default value to plot only positive non-zero values using color scale and zero/negative values as NA. To plot all values using color palette set to NA.

raster

Convert points to raster format. Default is NULL which will rasterize by default if greater than 200,000 cells.

split.by

Variable in @meta.data to split the plot by.

num_columns

Number of columns in plot layout.

slot

Which slot to pull expression data from? Default is "data".

...

Extra parameters passed to FeaturePlot.

na_value

color to use for points below lower limit.

Value

A ggplot object

Examples

if (FALSE) {
FeaturePlot_scCustom(seurat_object = object, features = "Cx3cr1", colors_use = viridis_plasma_dark_high,
na_color = "lightgray")
}