Iterate Cluster Highlight Plot

Iterate the create plots with cluster of interest highlighted across all cluster (active.idents) in given Seurat Object

Usage

Iterate_Cluster_Highlight_Plot(
  seurat_object,
  highlight_color = "navy",
  background_color = "lightgray",
  pt.size = NULL,
  reduction = NULL,
  file_path = NULL,
  file_name = NULL,
  file_type = NULL,
  single_pdf = FALSE,
  dpi = 600,
  raster = NULL,
  ...
)

Arguments

seurat_object

Seurat object name.

highlight_color

Color to highlight cells (default "navy"). Can provide either single color to use for all clusters/plots or a vector of colors equal to the number of clusters to use (in order) for the clusters/plots.

background_color

non-highlighted cell colors.

pt.size

point size for both highlighted cluster and background.

reduction

Dimensionality Reduction to use (if NULL then defaults to Object default).

file_path

directory file path and/or file name prefix. Defaults to current wd.

file_name

name suffix to append after sample name.

file_type

File type to save output as. Must be one of following: ".pdf", ".png", ".tiff", ".jpeg", or ".svg".

single_pdf

saves all plots to single PDF file (default = FALSE). `file_type`` must be .pdf

dpi

dpi for image saving.

raster

Convert points to raster format. Default is NULL which will rasterize by default if greater than 200,000 cells.

...

Extra parameters passed toDimPlot.

Value

Saved plots

Examples

if (FALSE) {
Iterate_Cluster_Highlight_Plot(seurat_object = object, highlight_color = "navy", background_color = "lightgray",
file_path = "path/", file_name = "name", file_type = "pdf", single_pdf = TRUE)
}