Iterative Plotting of Gene Lists using Custom FeaturePlots

Create and Save plots for Gene list with Single Command

Iterate_FeaturePlot_scCustom(
  seurat_object,
  gene_list,
  colors_use = viridis_plasma_dark_high,
  na_color = "lightgray",
  na_cutoff = 1e-09,
  split.by = NULL,
  order = TRUE,
  return_plots = FALSE,
  file_path = NULL,
  file_name = NULL,
  file_type = NULL,
  single_pdf = FALSE,
  dpi = 600,
  pt.size = NULL,
  reduction = NULL,
  raster = NULL,
  alpha_exp = NULL,
  alpha_na_exp = NULL,
  ...
)

Arguments

seurat_object

Seurat object name.

gene_list

vector of genes to plot. If a named vector is provided then the names for each gene will be incorporated into plot title if single_pdf = TRUE or into file name if FALSE.

colors_use

color scheme to use.

na_color

color for non-expressed cells.

na_cutoff

Value to use as minimum expression cutoff. To set no cutoff set to NA.

split.by

Variable in @meta.data to split the plot by.

order

whether to move positive cells to the top (default = TRUE).

return_plots

logical. Whether to return plots to list instead of saving them to file(s). Default is FALSE.

file_path

directory file path and/or file name prefix. Defaults to current wd.

file_name

name suffix and file extension.

file_type

File type to save output as. Must be one of following: ".pdf", ".png", ".tiff", ".jpeg", or ".svg".

single_pdf

saves all plots to single PDF file (default = FALSE). `file_type`` must be .pdf.

dpi

dpi for image saving.

pt.size

Adjust point size for plotting.

reduction

Dimensionality Reduction to use (if NULL then defaults to Object default).

raster

Convert points to raster format. Default is NULL which will rasterize by default if greater than 200,000 cells.

alpha_exp

new alpha level to apply to expressing cell color palette (colors_use). Must be value between 0-1.

alpha_na_exp

new alpha level to apply to non-expressing cell color palette (na_color). Must be value between 0-1.

...

Extra parameters passed to FeaturePlot.

Value

Saved plots

Examples

if (FALSE) {
Iterate_FeaturePlot_scCustom(seurat_object = object, gene_list = DEG_list,
colors_use = viridis_plasma_dark_high, na_color = "lightgray", file_path = "plots/",
file_name = "tsne", file_type = ".jpg", dpi = 600)
}