Iterative Plotting of Gene Lists using Custom Joint Density Plots

Create and save plots for gene list with single command. Requires Nebulosa package from Bioconductor.

Iterate_Plot_Density_Joint(
  seurat_object,
  gene_list,
  viridis_palette = "magma",
  custom_palette = NULL,
  pt.size = 1,
  file_path = NULL,
  file_name = NULL,
  file_type = NULL,
  single_pdf = FALSE,
  dpi = 600,
  reduction = NULL,
  combine = TRUE,
  joint = NULL,
  ...
)

Arguments

seurat_object

Seurat object name.

gene_list

a list of vectors of genes to plot jointly. Each entry in the list will be plotted for the joint density. All entries in list must be greater than 2 features. If a named list is provided then the names for each list entry will be incorporated into plot title if single_pdf = TRUE or into file name if FALSE.

viridis_palette

color scheme to use.

custom_palette

color for non-expressed cells. #' @param pt.size Adjust point size for plotting.

file_name

name suffix and file extension.

file_type

File type to save output as. Must be one of following: ".pdf", ".png", ".tiff", ".jpeg", or ".svg".

single_pdf

saves all plots to single PDF file (default = FALSE). `file_type`` must be .pdf.

dpi

dpi for image saving.

reduction

Dimensionality Reduction to use (if NULL then defaults to Object default) #' @param joint NULL. This function only supports joint = FALSE. Leave as NULL to generate plots. To iterate joint plots see function: Iterate_Plot_Density_Joint.

combine

Create a single plot? If FALSE, a list with ggplot objects is returned.

...

Extra parameters passed to plot_density.

file_path/prefix

directory file path and/or file name prefix. Defaults to current wd.

Examples

if (FALSE) {
Iterate_Plot_Density_Joint(seurat_object = object, gene_list = DEG_list, viridis_palette = "magma",
file_path = "plots/", file_name = "joint_plots", file_type = ".jpg", dpi = 600)
}