Plot Median other variable per Cell per Sample

Plot of median other variable per cell per sample grouped by desired meta data variable.

Usage

Plot_Median_Other(
  seurat_object,
  median_var,
  sample_col = "orig.ident",
  group_by = NULL,
  colors_use = NULL,
  plot_title = NULL,
  y_axis_label = NULL,
  x_axis_label = NULL,
  legend_title = NULL,
  x_lab_rotate = TRUE,
  color_seed = 123
)

Arguments

seurat_object

Seurat object name.

median_var

Variable in meta.data slot to calculate and plot median values for.

sample_col

Specify which column in meta.data specifies sample ID (i.e. orig.ident).

group_by

Column in meta.data slot to group results by (i.e. "Treatment").

colors_use

List of colors or color palette to use. Only applicable if group_by is not NULL.

plot_title

Plot title.

y_axis_label

Label for y axis.

x_axis_label

Label for x axis.

legend_title

Label for plot legend.

x_lab_rotate

logical. Whether to rotate the axes labels on the x-axis. Default is FALSE.

color_seed

random seed for the "varibow" palette shuffle if colors_use = NULL and number of groups plotted is greater than 36. Default = 123.

Value

A ggplot object

Examples

if (FALSE) {
Plot_Median_Other(seurat_object = obj, median_var = "module_score", sample_col = "orig.ident", group_by = "Treatment")
}