QC Plots Mito

#' Custom VlnPlot for initial QC checks including lines for thresholding

QC_Plots_Mito(
  seurat_object,
  mito_name = "percent_mito",
  plot_title = "Mito Gene % per Cell/Nucleus",
  group.by = NULL,
  x_axis_label = NULL,
  y_axis_label = "% Mitochondrial Gene Counts",
  low_cutoff = NULL,
  high_cutoff = NULL,
  pt.size = NULL,
  colors_use = NULL,
  x_lab_rotate = TRUE,
  y_axis_log = FALSE,
  raster = NULL,
  ggplot_default_colors = FALSE,
  color_seed = 123,
  ...
)

Arguments

seurat_object

Seurat object name.

mito_name

The column name containing percent mitochondrial counts information. Default value is "percent_mito" which is default value created when using Add_Mito_Ribo_Seurat().

plot_title

Plot Title.

group.by

Name of one or more metadata columns to group (color) cells by (for example, orig.ident); default is the current active.ident of the object.

x_axis_label

Label for x axis.

y_axis_label

Label for y axis.

low_cutoff

Plot line a potential low threshold for filtering.

high_cutoff

Plot line a potential high threshold for filtering.

pt.size

Point size for plotting

colors_use

vector of colors to use for plot.

x_lab_rotate

Rotate x-axis labels 45 degrees (Default is TRUE).

y_axis_log

logical. Whether to change y axis to log10 scale (Default is FALSE).

raster

Convert points to raster format. Default is NULL which will rasterize by default if greater than 100,000 total points plotted (# Cells x # of features).

ggplot_default_colors

logical. If colors_use = NULL, Whether or not to return plot using default ggplot2 "hue" palette instead of default "polychrome" or "varibow" palettes.

color_seed

random seed for the "varibow" palette shuffle if colors_use = NULL and number of groups plotted is greater than 36. Default = 123.

...

Extra parameters passed to VlnPlot.

Value

A ggplot object

Examples

if (FALSE) {
QC_Plots_Mito(seurat_object = object, plot_title = "Percent Mito per Cell", high_cutoff = 10)
}