Rename Cluster Seurat

Wrapper function to rename active identities in Seurat Object with new idents.

Usage

Rename_Clusters(seurat_object, new_idents, meta_col_name = NULL, ...)

Arguments

seurat_object

object name.

new_idents

vector of new cluster names. Must be equal to the length of current active.ident in Seurat Object. Will accept named vector (with old idents as names) or will name the new_idents vector internally.

meta_col_name

(Optional). Whether or not to create new named column in Object@meta.data to store the old identities.

...

Extra parameters passed to RenameIdents.

Value

Seurat Object with new identities placed in active.ident slot.

Examples

if (FALSE) {
obj <- Rename_Clusters(seurat_object = obj_name, new_idents = new_idents_vec, meta_col_name = "Round01_Res0.6_Idents")
}