QC Plots Sequencing metrics (Alignment)

Plot the fraction of reads confidently mapped to intergenic regions

Seq_QC_Plot_Intergenic(
  metrics_dataframe,
  plot_by = "sample_id",
  colors_use = NULL,
  x_lab_rotate = FALSE,
  significance = FALSE,
  ...
)

Arguments

metrics_dataframe

data.frame contain Cell Ranger QC Metrics (see Read_Metrics_10X).

plot_by

Grouping factor for the plot. Default is to plot as single group with single point per sample.

colors_use

colors to use for plot if plotting by group. Defaults to RColorBrewer Dark2 palette if less than 8 groups and DiscretePalette_scCustomize(palette = "polychrome") if more than 8.

x_lab_rotate

logical. Whether to rotate the axes labels on the x-axis. Default is FALSE.

significance

logical. Whether to calculate and plot p-value comparisons when plotting by grouping factor. Default is FALSE.

...

Other variables to pass to ggpubr::stat_compare_means when doing significance testing.

Value

A ggplot object

Examples

if (FALSE) {
Seq_QC_Plot_Intergeneic(metrics_dataframe = metrics)
}