Split FeatureScatter

Create FeatureScatter using split.by

Split_FeatureScatter(
  seurat_object,
  feature1 = NULL,
  feature2 = NULL,
  split.by = NULL,
  group.by = NULL,
  colors_use = NULL,
  pt.size = NULL,
  title_size = 15,
  num_columns = NULL,
  raster = NULL,
  raster.dpi = c(512, 512),
  ggplot_default_colors = FALSE,
  color_seed = 123,
  ...
)

Arguments

seurat_object

Seurat object name.

feature1

First feature to plot.

feature2

Second feature to plot.

split.by

Feature to split plots by (i.e. "orig.ident").

group.by

Name of one or more metadata columns to group (color) cells by (for example, orig.ident). Use 'ident' to group.by active.ident class.

colors_use

color for the points on plot.

pt.size

Adjust point size for plotting.

title_size

size for plot title labels.

num_columns

number of columns in final layout plot.

raster

Convert points to raster format. Default is NULL which will rasterize by default if greater than 100,000 cells.

raster.dpi

Pixel resolution for rasterized plots, passed to geom_scattermore(). Default is c(512, 512).

ggplot_default_colors

logical. If colors_use = NULL, Whether or not to return plot using default ggplot2 "hue" palette instead of default "polychrome" or "varibow" palettes.

color_seed

random seed for the "varibow" palette shuffle if colors_use = NULL and number of groups plotted is greater than 36. Default = 123.

...

Extra parameters passed to FeatureScatter.

Value

A ggplot object

Examples

if (FALSE) {
Split_FeatureScatter(seurat_object = object, features1= "nCount_RNA", feature2 = "nFeature_RNA",
split.by = "orig.ident", colors_use = "navy")
}