Store misc data in Seurat object

Wrapper function save variety of data types to the object@misc slot of Seurat object.

Store_Misc_Info_Seurat(
  seurat_object,
  data_to_store,
  data_name,
  list_as_list = FALSE,
  overwrite = FALSE,
  verbose = TRUE
)

Arguments

seurat_object

object name.

data_to_store

data to be stored in @misc slot. Can be single piece of data or list. If list of data see list_as_list parameter for control over data storage.

data_name

name to give the entry in @misc slot. Must be of equal length of the number of data items being stored.

list_as_list

logical. If data_to_store is a list, this dictates whether to store in @misc slot as list (TRUE) or whether to store each entry in the list separately (FALSE). Default is FALSE.

overwrite

Logical. Whether to overwrite existing items with the same name. Default is FALSE, meaning that function will abort if item with data_name is present in misc slot.

verbose

logical, whether to print messages when running function, default is TRUE.

Value

Seurat Object with new entries in the @misc slot.

Examples

library(Seurat)
clu_pal <- c("red", "green", "blue")

pbmc_small <- Store_Misc_Info_Seurat(seurat_object = pbmc_small, data_to_store = clu_pal,
data_name = "rd1_colors")
#> Seurat Object now contains the following items in @misc slot:
#> ℹ 'rd1_colors'