Calculate percent of expressing cells

Calculates the percent of cells that express a given set of features by various grouping factors

Percent_Expressing(
  seurat_object,
  features,
  threshold = 0,
  group_by = NULL,
  split_by = NULL,
  entire_object = FALSE,
  slot = deprecated(),
  layer = "data",
  assay = NULL
)

Arguments

seurat_object

Seurat object name.

features

Feature(s) to plot.

threshold

Expression threshold to use for calculation of percent expressing (default is 0).

group_by

Factor to group the cells by.

split_by

Factor to split the groups by.

entire_object

logical (default = FALSE). Whether to calculate percent of expressing cells across the entire object as opposed to by cluster or by group_by variable.

slot

soft-deprecated. See layer

layer

Which layer to pull expression data from? Default is "data".

assay

Assay to pull feature data from. Default is active assay.

Value

A data.frame

References

Part of code is modified from Seurat package as used by DotPlot to generate values to use for plotting. Source code can be found here: https://github.com/satijalab/seurat/blob/4e868fcde49dc0a3df47f94f5fb54a421bfdf7bc/R/visualization.R#L3391 (License: GPL-3).

Examples

if (FALSE) {
percent_stats <- Percent_Expressing(seurat_object = object, features = "Cx3cr1", threshold = 0)
}