QC Histogram Plots

Custom histogram for initial QC checks including lines for thresholding

QC_Histogram(
  seurat_object,
  features,
  low_cutoff = NULL,
  high_cutoff = NULL,
  cutoff_line_width = NULL,
  split.by = NULL,
  bins = 250,
  colors_use = "dodgerblue",
  num_columns = NULL,
  plot_title = NULL,
  assay = NULL,
  print_defaults = FALSE
)

Arguments

seurat_object

Seurat object name.

features

Feature from meta.data, assay features, or feature name shortcut to plot.

low_cutoff

Plot line a potential low threshold for filtering.

high_cutoff

Plot line a potential high threshold for filtering.

cutoff_line_width

numerical value for thickness of cutoff lines, default is NULL.

split.by

Feature to split plots by (i.e. "orig.ident").

bins

number of bins to plot default is 250.

colors_use

color to fill histogram bars, default is "dodgerblue".

num_columns

Number of columns in plot layout.

plot_title

optional, vector to use for plot title. Default is the name of the variable being plotted.

assay

assay to pull features from, default is active assay.

print_defaults

return list of accepted default shortcuts to provide to features instead of full name.

Value

A patchwork object

Examples

if (FALSE) {
QC_Histogram(seurat_object = object, features = "nFeature_RNA")
}