scCustomize: Custom Visualizations & Functions for Streamlined Analyses of Single Cell Sequencing

Collection of functions created and/or curated to aid in the visualization and analysis of single-cell data using 'R'. 'scCustomize' aims to provide 1) Customized visualizations for aid in ease of use and to create more aesthetic and functional visuals. 2) Improve speed/reproducibility of common tasks/pieces of code in scRNA-seq analysis with a single or group of functions. For citation please use: Marsh SE (2021) "Custom Visualizations & Functions for Streamlined Analyses of Single Cell Sequencing" doi:10.5281/zenodo.5706430 RRID:SCR_024675.

Package options

scCustomize uses the following options() to configure behavior:

scCustomize_warn_raster_iterative

Show message about setting raster parameter in Iterate_FeaturePlot_scCustom if raster = FALSE and single_pdf = TRUE due to large file sizes.

scCustomize_warn_raster_LIGER

Show warning about rasterization of points in DimPlot_LIGER due to new functionality compared to LIGER.

scCustomize_warn_na_cutoff

Show message about properly setting na_cutoff parameter in FeaturePlot_scCustom.

#'
scCustomize_warn_zero_na_cutoff

Show message about properly setting na_cutoff parameter in FeaturePlot_scCustom if na_cutoff is set to exactly zero.

scCustomize_warn_vln_raster_iterative

Show message about Iterate_VlnPlot_scCustom when pt.size > 0 due to current lack of raster support in VlnPlot

scCustomize_warn_LIGER_dim_labels

Show message about DimPlot_LIGER parameter reduction_label as LIGER objects do not store dimensionality reduction name and and therefore needs to be set manually.

scCustomize_warn_DimPlot_split_type

Show message about DimPlot_scCustom parameter split.by and split_seurat to alert user to difference in returned plots between scCustomize and Seurat.

scCustomize_warn_FeatureScatter_split_type

Show message about FeatureScatter_scCustom parameter split.by and split_seurat to alert user to difference in returned plots between scCustomize and Seurat.

scCustomize_warn_LIGER_dim_labels_plotFactors

Show message about plotFactors_scCustom parameter reduction_label as LIGER objects do not store dimensionality reduction name and and therefore needs to be set manually.

See also

Author

Maintainer: Samuel Marsh sccustomize@gmail.com (ORCID)

Other contributors:

Ming Tang tangming2005@gmail.com [contributor]

Velina Kozareva [contributor]

Lucas Graybuck lucasg@alleninstitute.org [contributor]

Zoe Clarke zoe.clarke@utoronto.ca (ORCID) [contributor]